+++ CHANGELOG.md
@@ -6,7 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
-
+## [0.1.2] - 2019-05-13
+### Added
+- delegators in transaction and pledge
+- statuses methods in transaction
 ## [0.1.0] - 2019-05-05
 ### Added
 - Created darujme_cz prototype